# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiErrorResponse": {
        "description": "统一错误响应包。任何非 2xx 的应用错误都使用该结构。",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "DraftPatchRequest": {
        "additionalProperties": false,
        "description": "`PATCH /items/{id}/drafts/{draftId}` 请求体（T19 受限字段集）。\n\n每个字段都可缺省；**至少提供一个字段**，否则 422（空请求体）。",
        "properties": {
          "clearStepPoses": {
            "description": "要清除视角的步骤 id（「清除该步骤视角」动作；与写入分开）。",
            "items": {
              "type": "string"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "entities": {
            "additionalProperties": {
              "$ref": "#/components/schemas/EntityReviewPatchDto"
            },
            "description": "实体级复核（确认/取消确认、人工修订、仅文本条目）。",
            "propertyNames": {
              "type": "string"
            },
            "type": [
              "object",
              "null"
            ]
          },
          "hotspots": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/HotspotPatch",
                "description": "热点写入（新建/重新绑定/解绑；人工直接拾取 = upsert confirmed + anchor）。"
              }
            ]
          },
          "interactive": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/InteractivePatch",
                "description": "交互层（ADR-042）：部件→分件节点绑定、动作、姿势（整体替换给出的集合）。"
              }
            ]
          },
          "modelReview": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ModelReviewPatch",
                "description": "modelReview 的两个用户声明（`checkedAt` 由服务器赋值）。"
              }
            ]
          },
          "status": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DraftStatusDto",
                "description": "目标状态（`needs_review` / `ready`）；不触发发布。"
              }
            ]
          },
          "stepPoses": {
            "additionalProperties": {
              "$ref": "#/components/schemas/CameraPose"
            },
            "description": "步骤视角写入：`{ \"<stepId>\": CameraPose }`（保存/覆盖；「保存当前视角」动作）。",
            "propertyNames": {
              "type": "string"
            },
            "type": [
              "object",
              "null"
            ]
          }
        },
        "type": "object"
      },
      "DraftResponse": {
        "description": "`GET /items/{id}/drafts/{draftId}` 响应（带 `ETag`）。",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/DraftDto"
          }
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
      "DraftStatusDto": {
        "description": "草稿状态（线上 snake_case：`needs_review` / `ready`）。",
        "enum": [
          "needs_review",
          "ready"
        ],
        "type": "string"
      },
...
      "ReleaseResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ReleaseDto"
          }
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
...
      "SessionResponse": {
        "description": "`GET /api/v1/auth/session` 响应（`200`；`Cache-Control: no-store`）。",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/SessionData"
          }
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "sessionCookie": {
        "in": "cookie",
        "name": "em_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "自托管交互说明书的本地 HTTP API。健康探针、认证/会话、设置状态、资产上传/内容服务（T06）、物品与说明书绑定/多视图照片（T07）、其余路由按 llmdoc/contracts.md 在同一文档上扩展。",
    "title": "万物说明书 API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/session": {
      "get": {
        "description": "返回管理员摘要与 CSRF token；响应带 Cache-Control: no-store。未登录返回 401。",
        "operationId": "session",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "当前会话"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "未登录或会话已失效"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "读取当前会话",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/items/{id}/drafts/{draftId}": {
      "get": {
        "description": "返回草稿的知识聚合（含 `completeness`/`missing[]`：部分成功可展示）、模型版本引用与复核状态，带 `ETag: \"r<revision>\"` 供 PATCH 的 If-Match。`status = needs_review` 只表示可复核草稿已产出——生成完成不等于已发布，不存在自动发布路径。",
        "operationId": "get_draft",
        "parameters": [
          {
            "description": "物品 ID（UUIDv7）",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "草稿 ID（UUIDv7）",
            "in": "path",
            "name": "draftId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftResponse"
                }
              }
            },
            "description": "草稿（带 ETag）"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "未登录"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "草稿不存在或不属于该物品"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "读取版本化草稿（带 ETag）",
        "tags": [
          "drafts"
        ]
      },
      "patch": {
        "description": "必须携带 If-Match（来自 GET 的 ETag）：缺失 428、revision 过期 412。受限字段：`status`（needs_review ↔ ready）、`hotspots`（新建/重新绑定/解绑；人工直接拾取 = upsert confirmed + anchor；unbound 时 anchor 必须为 null，不得用 [0,0,0] 占位；candidate/confirmed 的 anchor 必须与当前模型 revision+sha 完全一致，旧 sha 一律 422）、`stepPoses`（CameraPose 或 null 清除；数值有限、fov 合理范围）、`entities`（confirmed/needs_review 切换、人工修订 userEdited、仅文本条目 textOnly）、`modelReview`（loaded 与 userConfirmed 两个用户声明，checkedAt 由服务器赋值）。**不能修改供应商事实快照**：knowledge 里的 Part/Step/Evidence 只读，人工修订写入独立覆盖层并保留原文本与出处；未知字段 422。空请求体 422；无实际变化的重复提交按幂等返回（不递增 revision，stale 仍 412）。动作写入 audit_events。",
        "operationId": "patch_draft",
        "parameters": [
          {
            "description": "物品 ID（UUIDv7）",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "草稿 ID（UUIDv7）",
            "in": "path",
            "name": "draftId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftPatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftResponse"
                }
              }
            },
            "description": "更新后的草稿（带新 ETag）"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "未登录"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "CSRF/Origin 校验失败"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "草稿不存在或不属于该物品"
          },
          "412": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "revision 过期（details.currentRevision）"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "字段级校验失败（details.fields；含未知字段/引用不存在/旧 sha 拒绝）"
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "缺少 If-Match"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "复核写入：状态、热点校准、步骤视角、知识确认与修订、modelReview（If-Match）",
        "tags": [
          "drafts"
        ]
      }
    },
    "/api/v1/items/{id}/drafts/{draftId}/parts-model": {
      "post": {
        "description": "multipart：`file`（GLB，必填）+ `source`（来源说明，可选）。服务端校验 GLB 预算、节点名唯一且只含平移，并核对分件包围盒与草稿模型 `bounds` 一致（坐标系相同）；通过后写入 `knowledge.interactive.partsModel`，旧绑定/动作中引用已不存在节点的条目被移除。不改变模型版本与热点锚点身份；动作写入 audit_events。",
        "operationId": "attach_parts_model",
        "parameters": [
          {
            "description": "物品 ID（UUIDv7）",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "草稿 ID（UUIDv7）",
            "in": "path",
            "name": "draftId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftResponse"
                }
              }
            },
            "description": "已挂载（返回更新后的草稿）"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "未登录"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "CSRF/Origin 校验失败"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "草稿不存在或不属于该物品"
          },
          "412": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "revision 过期"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "文件超过模型体积上限"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "不是合法分件 GLB 或坐标系不一致（details.fields）"
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "缺少 If-Match"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "挂载分件模型（与草稿模型同一坐标系的多节点 GLB；If-Match）",
        "tags": [
          "drafts"
        ]
      }
    },
    "/api/v1/items/{id}/drafts/{draftId}/publish": {
      "post": {
        "description": "发布事务在**显式请求**中执行（不存在自动发布路径）。必须携带 If-Match （缺失 428、过期 412）与 `Idempotency-Key`（缺失 422；同 key 同 body 重放返回同一 release 并标记 `x-idempotent-replay: true`；同 key 不同 body 409）。发布不变量全部满足才返回 201：必需知识已确认或有明确人工修订记录、引用页存在、选中模型 validated 且 modelReview.loaded/userConfirmed 均 true 且 revision/hash 匹配、每个要发布的交互部件至少一个 confirmed 热点且 hash 匹配、步骤引用全部存在、无 stale/candidate 热点冒充 confirmed。不满足返回 422 且 `details.issues[]` 逐条列出；「仅文本条目」部件保留并明显标识，不为发布自动隐藏。release 与 manifest 不可变（数据库触发器拒绝修改），发布只写本地数据——不产生费用、不调用任何外部服务。",
        "operationId": "publish_draft",
        "parameters": [
          {
            "description": "物品 ID（UUIDv7）",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "草稿 ID（UUIDv7）",
            "in": "path",
            "name": "draftId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseResponse"
                }
              }
            },
            "description": "已发布（不可变 release；重放同一 release）"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "未登录"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "CSRF/Origin 校验失败"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "草稿不存在或不属于该物品"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "同 key 不同 body（IDEMPOTENCY_CONFLICT）"
          },
          "412": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "revision 过期/并发修改（details.currentRevision）"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "缺 Idempotency-Key（details.fields）或发布不变量不满足（details.issues）"
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "缺少 If-Match"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "发布不可变说明书版本（If-Match + Idempotency-Key）",
        "tags": [
          "drafts"
        ]
      }
    },
...
  }
}